redesign

+++ kernel/vds_crypto/asymmetriccrypto_p.h
@@ -94,7 +94,7 @@ namespace vds {
 
     data_buffer encrypt(const data_buffer & data)
     {
-      int blocksize = RSA_size(EVP_PKEY_get1_RSA(this->key_)) - 11;// EVP_PKEY_size(this->key_);
+      size_t blocksize = (size_t)RSA_size(EVP_PKEY_get1_RSA(this->key_)) - 11;// EVP_PKEY_size(this->key_);
       data_buffer result;
 
       auto p = data.data();
